# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CarRental API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/payments": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Get list of payments",
                "operationId": "826dc6b24dd704607aa018a59207dd96",
                "parameters": [
                    {
                        "name": "rental_id",
                        "in": "query",
                        "description": "Filter by rental"
                    },
                    {
                        "name": "min_amount",
                        "in": "query",
                        "description": "Filter by min amount"
                    },
                    {
                        "name": "max_amount",
                        "in": "query",
                        "description": "Filter by max amount"
                    },
                    {
                        "name": "payment_date",
                        "in": "query",
                        "description": "Filter by payment date"
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of payments"
                    }
                }
            },
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Create a new payment",
                "operationId": "5e2794ba3d3c64e70873367505e4a751",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rental_id",
                                    "amount",
                                    "payment_date"
                                ],
                                "properties": {
                                    "rental_id": {
                                        "type": "integer"
                                    },
                                    "amount": {
                                        "type": "number"
                                    },
                                    "payment_date": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Payment created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/payments/{id}": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Get payment details",
                "operationId": "a9a7656e0f22931dbcedc57a2d959554",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Payment ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment details"
                    }
                }
            },
            "put": {
                "tags": [
                    "Payments"
                ],
                "summary": "Update payment details",
                "operationId": "9f607012ea3accda178adb489aa28da2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Payment ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rental_id",
                                    "amount",
                                    "payment_date"
                                ],
                                "properties": {
                                    "rental_id": {
                                        "type": "integer"
                                    },
                                    "amount": {
                                        "type": "number"
                                    },
                                    "payment_date": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Payments"
                ],
                "summary": "Delete a payment",
                "operationId": "eee625edf09d69ad0d35b893a3b4f087",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Payment ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}